Move InternalClient & rename it to NetClient
